helper in util
